* Resolve one bug related with size of transformation field * Resolve comments given by @stebogit
