+++ packages/turf-constrained-tin/test/in/mercator_1.json
