+++ packages/turf-constrained-tin/test/in/mercator_2.json
